More tests marked as ITs
