+++ extensions/mapstore/src/test/java/com/hazelcast/mapstore/mysql/MySQLGenericMapStoreBasicIT.java
@@ -14,34 +14,20 @@
  * limitations under the License.
  */
 
-package com.hazelcast.mapstore.oracle;
+package com.hazelcast.mapstore.mysql;
 
-import com.hazelcast.mapstore.GenericMapStoreTest;
+import com.hazelcast.mapstore.GenericMapStoreBasicIT;
 import com.hazelcast.test.annotation.NightlyTest;
-import com.hazelcast.test.jdbc.OracleDatabaseProvider;
+import com.hazelcast.test.jdbc.MySQLDatabaseProvider;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
 
-import java.sql.SQLException;
-
 @Category(NightlyTest.class)
-public class OracleGenericMapStoreTest extends GenericMapStoreTest {
+public class MySQLGenericMapStoreBasicIT extends GenericMapStoreBasicIT {
 
     @BeforeClass
     public static void beforeClass() {
-        initialize(new OracleDatabaseProvider());
+        initialize(new MySQLDatabaseProvider());
     }
 
-    @Override
-    protected void createMapLoaderTable(String tableName) throws SQLException {
-        createTable(tableName, "id NUMBER(8) PRIMARY KEY", "name VARCHAR(100)");
-    }
-
-    @Override
-    protected void createMapLoaderTable(String tableName, String... columns) throws SQLException {
-        for (int i = 0; i < columns.length; i++) {
-            columns[i] = columns[i].replace("INT", "NUMBER(8)");
-        }
-        createTable(tableName, columns);
-    }
 }
