+++ extensions/mapstore/src/test/java/com/hazelcast/mapstore/oracle/OracleGenericMapStoreBasicIT.java
@@ -14,20 +14,34 @@
  * limitations under the License.
  */
 
-package com.hazelcast.mapstore.mysql;
+package com.hazelcast.mapstore.oracle;
 
-import com.hazelcast.mapstore.GenericMapStoreTest;
+import com.hazelcast.mapstore.GenericMapStoreBasicIT;
 import com.hazelcast.test.annotation.NightlyTest;
-import com.hazelcast.test.jdbc.MySQLDatabaseProvider;
+import com.hazelcast.test.jdbc.OracleDatabaseProvider;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
 
+import java.sql.SQLException;
+
 @Category(NightlyTest.class)
-public class MySQLGenericMapStoreTest extends GenericMapStoreTest {
+public class OracleGenericMapStoreBasicIT extends GenericMapStoreBasicIT {
 
     @BeforeClass
     public static void beforeClass() {
-        initialize(new MySQLDatabaseProvider());
+        initialize(new OracleDatabaseProvider());
     }
 
+    @Override
+    protected void createMapLoaderTable(String tableName) throws SQLException {
+        createTable(tableName, "id NUMBER(8) PRIMARY KEY", "name VARCHAR(100)");
+    }
+
+    @Override
+    protected void createMapLoaderTable(String tableName, String... columns) throws SQLException {
+        for (int i = 0; i < columns.length; i++) {
+            columns[i] = columns[i].replace("INT", "NUMBER(8)");
+        }
+        createTable(tableName, columns);
+    }
 }
