+++ extensions/mapstore/src/test/java/com/hazelcast/mapstore/postgres/PostgresGenericMapStoreSimpleIT.java
@@ -16,14 +16,14 @@
 
 package com.hazelcast.mapstore.postgres;
 
-import com.hazelcast.mapstore.GenericMapStoreTest;
+import com.hazelcast.mapstore.GenericMapStoreBasicIT;
 import com.hazelcast.test.annotation.NightlyTest;
 import com.hazelcast.test.jdbc.PostgresDatabaseProvider;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
 
 @Category(NightlyTest.class)
-public class PostgresGenericMapStoreSimpleIT extends GenericMapStoreTest {
+public class PostgresGenericMapStoreSimpleIT extends GenericMapStoreBasicIT {
 
     @BeforeClass
     public static void beforeClass()  {
